Fix broken link
Support for Ubuntu 19.10 ("Eoan Ermine") ended in July 2020, and
it is no longer listed on the Ubuntu website when searching for
packages, so the link for eoan just shows an error. To fix this,
it is replaced by the corresponding link for Ubuntu 22.04 ("Jammy
Jellyfish"), the current LTS release supported until April 2027,
so the link should keep working for the next few years.